Migrate to vite and fix rendering
