+++ .gitignore
@@ -16,3 +16,6 @@ yarn-debug.log*
 yarn-error.log*
 
 .gitpod.yml
+
+dist
+.vscode
