+++ index.html
@@ -1,0 +1,21 @@
+<!DOCTYPE html>
+<html lang="en">
+  <head>
+    <meta charset="utf-8" />
+    <link rel="icon" href="/favicon.ico" />
+    <meta name="viewport" content="width=device-width, initial-scale=1" />
+    <meta name="theme-color" content="#000000" />
+    <meta
+      name="Pathfinder"
+      content="Pathfinder website simulating Algorithms between two nodes"
+    />
+    <link rel="apple-touch-icon" href="/logo192.png" />
+    <link rel="manifest" crossorigin="use-credentials" href="manifest.json" />
+    <title>Pathfinder Simulation</title>
+  </head>
+  <body>
+    <noscript>You need to enable JavaScript to run this app.</noscript>
+    <div id="root"></div>
+    <script type="module" src="/src/index.tsx"></script>
+  </body>
+</html>
