+++ script.sh
@@ -1,0 +1,1 @@
+npm i --save-dev @types/node
