--- src/index.js
@@ -1,9 +1,0 @@
-import React from 'react';
-import ReactDOM from 'react-dom';
-import './index.css';
-import App from './App';
-import * as serviceWorker from "./serviceWorker";
-
-ReactDOM.render(<App />, document.getComponent("root"));
-
-serviceWorker.unregister();
