+++ vite.config.ts
@@ -1,0 +1,8 @@
+import { defineConfig } from "vite";
+import react from "@vitejs/plugin-react";
+import viteTsconfigPaths from "vite-tsconfig-paths";
+import svgrPlugin from "vite-plugin-svgr";
+
+export default defineConfig({
+  plugins: [react(), viteTsconfigPaths(), svgrPlugin()],
+});
